# Fix a bug that BufferChunkOverflowError cause a whole data loss