**Which issue(s) this PR fixes**: 

Fixes #

**What this PR does / why we need it**: 

In the previous version, if BufferChunkOverflowError occurs,
whole data is not sent because of exception.
It is not appropriate because the problem exists only a partial split
data.

In this commit, detect BufferChunkOverflowError and postpone to raise
it as an exception.
So a valid partial split is processed as intended.

**Docs Changes**:

N/A

**Release Note**: 

N/A