@kenhys 2021-11-12
I've considered two-approaches to fix again.
* propagate error via #write_once block
* propagate error by raising #write_step_by_step

The first one is confirmed that it shows better performance.