@ashie 2021-11-16
@kenhys Could you take a look #1849 too?
Although it's a problem that is different from the one we are going to solve (whole size of a chunk vs 1 message's size), but a very near issue.