@kenhys 2021-11-16
With extending memory chunk approach (PoC), it seems that it overcome the performance regression:

```
 Before:

      Warming up --------------------------------------
      no BufferChunkOverflowError
                              11.488k i/100ms
      raise BufferChunkOverflowError in every 100 calls
                               4.110k i/100ms
      Calculating -------------------------------------
      no BufferChunkOverflowError
                              113.353k (± 1.0%) i/s -      1.137M in  10.034300s
      raise BufferChunkOverflowError in every 100 calls
                               40.599k (± 3.5%) i/s -    406.890k in  10.034076s

      Comparison:
      no BufferChunkOverflowError:   113353.4 i/s
      raise BufferChunkOverflowError in every 100 calls:    40599.5 i/s - 2.79x  (± 0.00) slower

    After:

      Warming up --------------------------------------
      no BufferChunkOverflowError
                              11.399k i/100ms
      raise BufferChunkOverflowError in every 100 calls
                               4.301k i/100ms
      Calculating -------------------------------------
      no BufferChunkOverflowError
                              113.855k (± 1.1%) i/s -      1.140M in  10.013090s
      raise BufferChunkOverflowError in every 100 calls
                               43.683k (± 4.0%) i/s -    438.702k in  10.059251s

      Comparison:
      no BufferChunkOverflowError:   113854.6 i/s
      raise BufferChunkOverflowError in every 100 calls:    43683.3 i/s - 2.61x  (± 0.00) slower

```