@kenhys 2021-11-16
TODO: also do the same thing for file chunk.