@ashie 2021-11-16
Looks better than before.
But I'm wondering it's really needed to modify chunk's code.
I think the size can be determined before calling `chunk.append` or `chunk.concat`.

https://github.com/fluent/fluentd/blob/2722b799e2aba3f82c279deb5d554fb44c077f4d/lib/fluent/plugin/buffer.rb#L774-L778

https://github.com/fluent/fluentd/blob/2722b799e2aba3f82c279deb5d554fb44c077f4d/lib/fluent/plugin/buffer/chunk.rb#L87-L99