@kenhys 2021-11-16
tried another approach without modifying chunk code: just calculate size in advance.
It also has no performance drawbacks.