@kenhys 2021-11-16
As for #1849, all 3 records exceeds the chunk_size_limit.

https://github.com/fluent/fluentd/issues/1849#issuecomment-968781444

```
2021-11-16 16:45:55 +0900 [warn]: #0 emit transaction failed: error_class=Fluent::Plugin::Buffer::BufferChunkOverflowError error="a 67 bytes record (nth: 0) is larger than buffer chunk limit size, a 67 bytes record (nth: 1) is larger than buffer chunk limit size, a 67 bytes record (nth: 2) is larger than buffer chunk limit size" location="/home/kenhys/.rvm/gems/ruby-2.7.4@fluentd-test/gems/fluentd-1.14.2/lib/fluent/plugin/buffer.rb:454:in `write'" tag="tail.log"
```