@ashie 2021-11-16
> As for #1849, all 3 records exceeds the chunk_size_limit.

Hmm, it's same problem with this?