@kenhys 2021-11-16
> Hmm, it's same problem with this?

I think so.

* https://github.com/fluent/fluentd/issues/1849#issue-295004393
  * indicates possibility to clean up overflow logic
* https://github.com/fluent/fluentd/issues/1849#issuecomment-968781444
  * point out data loss (same as #3553)

It can be improved overflow logic, butI'm not sure yet.