@ashie 2021-11-16
DCO check is failed: https://github.com/fluent/fluentd/pull/3553/checks?check_run_id=4221605400