@ashie 2021-11-17
> > As for #1849, all 3 records exceeds the chunk_size_limit.
> 
> Hmm, it's same problem with this?

No, it's a different problem.
For exmaple, in the case of https://github.com/fluent/fluentd/issues/1849#issuecomment-968781444, `BufferChunkOverflow` is occurred on writing 2nd record although all records don't exceed `chunk_size_limit`. They should be separated into different chunks.

We should work on it in another pull request.