@kenhys 2021-11-18
A bit strange because it pass on my local environment.

```
% bundle exec ruby test/plugin/test_buffer.rb 
Loaded suite test/plugin/test_buffer
Started
..............................................................
Finished in 0.579545554 seconds.
--------------------------------------------------------------------------------------------------------------------------------------------
62 tests, 376 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions, 0 notifications
100% passed
--------------------------------------------------------------------------------------------------------------------------------------------
106.98 tests/s, 648.78 assertions/s
```