@kenhys 2021-11-18
Even though `bundle exec rake test` on my local, it passes. (not depending the other test cases)