@kenhys 2021-11-18
```
      c = @p.generate_chunk(create_metadata)
      c.append(Fluent::ArrayEventStream.new([timestamp, {"message" => "0123"}]), compress: :gzip)
      p c.bytesize
```

it shows 49 when invoked from test case. it is expected 71