@kenhys 2021-11-18
> it shows 49 when invoked from test case. it is expected 71

digging internal flow, message "0123", "aaa", "bbb" are occupy 71, 70,70 bytes (chunk.bytesize)